feat(politics-tracker): add GA to all pages
