+++ packages/politics-tracker/components/layout/default.tsx
@@ -1,8 +1,7 @@
-import { gaTrackingId } from '~/constants/config'
-import Head from 'next/head'
 import Header from '~/components/header'
 import Footer from '~/components/footer'
 import ToastProvider from '~/components/toast/toast-provider'
+import GAScript from '~/components/ga-script'
 
 type DefaultLayoutProps = {
   children: React.ReactNode
@@ -13,24 +12,7 @@ export default function DefaultLayout({
 }: DefaultLayoutProps): JSX.Element {
   return (
     <>
-      <Head>
-        {/* ref: https://mariestarck.com/add-google-analytics-to-your-next-js-application-in-5-easy-steps/ */}
-        {/* Google tag (gtag.js) */}
-        <script
-          async
-          src={`https://www.googletagmanager.com/gtag/js?${gaTrackingId}`}
-        />
-        <script
-          dangerouslySetInnerHTML={{
-            __html: `
-              window.dataLayer = window.dataLayer || [];
-              function gtag(){dataLayer.push(arguments);}
-              gtag('js', new Date());
-              gtag('config', '${gaTrackingId}');
-            `,
-          }}
-        />
-      </Head>
+      <GAScript />
       <ToastProvider>
         <>
           <Header />
