+++ packages/politics-tracker/pages/person/[id].js
@@ -9,6 +9,7 @@ import Section from '~/components/person/section'
 import { cmsApiUrl } from '~/constants/config'
 import { print } from 'graphql'
 import { fireGqlRequest } from '~/utils/utils'
+import GAScript from '~/components/ga-script'
 import GetPersonBasicInfo from '~/graphql/query/person/get-person-basic-info.graphql'
 import GetPersonElections from '~/graphql/query/person/get-person-elections.graphql'
 const Main = styled.main`
@@ -26,6 +27,7 @@ const Main = styled.main`
 export default function People({ personData, personElectionsData }) {
   return (
     <ThemeProvider theme={theme}>
+      <GAScript />
       <Main>
         <Title name={personData.name} image={personData.image} />
         <Section
